pyproject.toml
# pyproject.toml
[project]
name = "notx-eval"
version = "0.1.0"
description = "Parallel longform generation + not-x-but-y slop evaluation"
requires-python = ">=3.11"
authors = [{ name = "Your Name" }]
readme = "README.md"
license = { text = "MIT" }

dependencies = [
  "requests>=2.32.0",
  "python-dotenv>=1.0.1",
  "tqdm>=4.66.0",
  # POS tagging and sentencizer. If you won't use Stage-2, you can omit spaCy.
  "spacy>=3.7.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5.0",
  "pytest>=8.2.0",
]

[project.scripts]
notx-eval = "main:main"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E","F","I","UP"]
fix = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
